+++ packages/modal/src/components/header/Component.tsx
@@ -44,8 +44,11 @@ export type HeaderProps = {
     /**
      * Выравнивание заголовка
      */
-    align?: 'left' | 'right' | 'center';
+    align?: 'left' | 'center';
 
+    /**
+     * Обрезать ли заголовок
+     */
     trim?: boolean;
 
     /**
@@ -95,7 +98,9 @@ export const Header: FC<HeaderProps> = ({
                 </div>
             )}
 
-            {closer && <div className={cn(styles.addon, addonClassName)}>{closer}</div>}
+            {closer && (
+                <div className={cn(styles.addon, styles.closer, addonClassName)}>{closer}</div>
+            )}
         </div>
     );
 };